Remove thanks-shim
